Group dialogs under DelegationDialog
